+++ src/modules/underground/UndergroundItemNameType.ts
@@ -4,7 +4,7 @@ To update this type when adding new items:
 Open the game, and run the following code in the browser console
 
 copy(`type UndergroundItemNameType
-    = ${[...new Set(Object.values(UndergroundItem.list).map(i => i.name))].map(i => `'${i.replace(/'/g, "\\'")}'`).join('\n    | ')};`);
+    = ${[...new Set(Object.values(UndergroundItems.list).map(i => i.name))].map(i => `'${i.replace(/'/g, "\\'")}'`).join('\n    | ')};`);
 
 Replace the everything in this file (except for this comment) with what was copied
 */
@@ -78,3 +78,5 @@ type UndergroundItemNameType
     | 'Cyan Shard'
     | 'Rose Shard'
     | 'Brown Shard';
+
+    export default UndergroundItemNameType;
